Move Carto::Http::ResponseLogger to separate file #3778
